Introduces new reshare interface to allow reshare of private posts
Resharing now brings up a special ACL selector.

If the post is public the member can simply click "Share" on the
selector and be done, no edit box involved. He can also make his
reshared post private if he wishes.

If the original is private the member must explicitly select an
audience, raising his awareness of the privacy issue and
avoiding unintentionally resharing to the default audience.

Private posts also have an option to 'hide the origin', which
removes identifying information from the header. We plan to deal
with included media later on, either by copying or removing it.

In all cases an "Edit" button is available to let the member
edit the message before resharing, much like how it was before.

It's a bit hacky until we know if and how it will get into main.